pyproject.toml
[project]
name = "platform_service_example"
version = "0.1.0"
description = "description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-ansible-base",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "poethepoet>=0.37.0",
    "ruff>=0.14.3",
    "ty>=0.0.1a25",
    "pdoc>=16.0.0",
    "django-extensions>=4.1",
    "ipython>=9.7.0",
    "ipdb>=0.13.13",
    "django-debug-toolbar>=6.1.0",
]
doc = [
    "pdoc>=16.0.0",
]
test = [
    "pytest>=8.4.2",
    "poethepoet>=0.37.0",
    "ruff>=0.14.3",
    "ty>=0.0.1a25",
]

[tool.ruff]
line-length = 100
exclude = ["__pycache__", ".venv", "build", "dist", ".git"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "N", "I"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-vs -ra -q"
testpaths = ["platform_service_example/tests", "apps/**/tests"]

[tool.poe]
executor.type  = "uv"

[tool.poe.tasks]
lint = "ruff check ."
type_check = "ty check"
format = "ruff format ."
unit-test = "manage.py test -v 2"
check = ["format", "lint","type_check", "unit-test"]
render-docs = "pdoc -t docs/templates --mermaid --docformat markdown --docformat google platform_service_example apps -o html"
serve-docs = "pdoc -t docs/templates --mermaid --docformat markdown --docformat google platform_service_example apps"

[tool.poe.tasks.clean]
help = "Clean local caches for mypy, pytest, python"
script = "poethepoet.scripts:rm('.mypy_cache', '.pytest_cache', './**/__pycache__', './**/*.pyc')"

[tool.poe.tasks.validate]
shell = """
python -c "
import yaml
with open('.copier-answers.yml') as f:
    d = yaml.safe_load(f)
    branch = d.get('src_branch', '')
    src = d.get('_src_path', '')
    from pathlib import Path
    if Path(src).is_dir():
        print(f'uvx --refresh --from {src} platform-service-framework validate')
    else:
        print(f'uvx --refresh git+{src}@{branch} validate')
" | sh
"""

[tool.poe.tasks.update]
shell = """
python -c "
import yaml
with open('.copier-answers.yml') as f:
    d = yaml.safe_load(f)
    branch = d.get('src_branch', '')
    src = d.get('_src_path', '')
    from pathlib import Path
    if Path(src).is_dir():
        print(f'uvx --refresh --from {src} platform-service-framework update')
    else:
        print(f'uvx --refresh git+{src}@{branch} update')
" | sh
"""

[tool.uv.sources]
django-ansible-base = { git = "https://github.com/ansible/django-ansible-base", rev = "devel" }
